int -> float on change duty cycle

+++ lab4-bg.py
@@ -33,7 +33,7 @@ while True:
     data = json.load(f)
 
     LED = int(data["LED"])
-    dutyCycle = int(data["Brightness"])
+    dutyCycle = float(data["Brightness"])
 
     # Old code, still works for troubleshooting but using json formatting instead
     # settings = f.read().split(',')  # create an array [ LED_selection , brighntess]
